requirements.txt
pytest-describe
mockito
responses
requests